wrap dialog manager in pinia store
