+++ src/modules/feature/dialog/index.ts
@@ -1,4 +1,4 @@
 export * from "./dialog-contracts";
-export { openDialog } from "./dialog-manager";
+export { openDialog } from "./dialog-manager.store";
 export type { AwaitableResult } from "./dialog-types";
 export { default as DialogHost } from "./DialogHost.vue";
